Complete Age Groups

Completes age groups by adding missing age groups, either within the age range from age_gt to age_lt or using the full age list from seer_std_ages. If the columns age_min or age_max are missing from the input data, separate_age_groups() is first called to expand the age group variable.

complete_age_groups(data, age_gt = NULL, age_lt = NULL, ...,
  age_group = age_group, fill = list(n = 0), include_unknown = FALSE,
  std_age_groups = fcds_const("age_group"))

Arguments

data
A data frame.

age_gt
Youngest age (inclusive).

age_lt
Eldest age (inclusive).

...
Not used other than to require explicit naming of arguments.

age_group
Unquoted column name containing the age grouping.

fill
A named list that for each variable supplies a single value to use instead of NA for missing combinations.

include_unknown
Should the "Unknown" age group be included?

std_age_groups
Character vector containing expected (or standard) age groups.

See also

Examples

dplyr::tibble(
  age_group = c("10 - 14", "15 - 19", "25 - 29"),
  n = 10:12
) %>%
  complete_age_groups(10, 35)

#> # A tibble: 5 x 2
#>   age_group     n
#>   <ord>     <dbl>
#> 1 10 - 14      10
#> 2 15 - 19      11
#> 3 20 - 24       0
#> 4 25 - 29      12
#> 5 30 - 34       0