Complete Year Groups

Completes year groups in the data frame with the expected year values, see fcds_const().

complete_year_groups(data, ..., year_min = NULL, year_max = NULL,
  year = year, fill = list(n = 0))

Arguments

data
A data frame

...
Ignored if data is a grouped data frame. If not grouped, additional arguments are passed to tidyr::complete(). Use these arguments specify which columns are included in the expansion and how. See tidyr::complete() for more information.

year_min
Optional earliest year to include (inclusive)

year_max
Optional latest year to include (inclusive)

year
The column containing the year_group.

fill
Default values for rows in columns added to the data

See also

Examples

d_year <- tidyr::crossing(
  sex = "Female",
  race = fcds_const("race"),
  year = fcds_const("year")
)

# These two versions are equivalent. The first version completes all variables
# included in the grouping and the second explicitly declares the variables
# that should be completed.

d_year %>%
  dplyr::group_by(sex, race) %>%
  complete_year_groups() %>%
  dplyr::arrange(sex, race, year)

#> # A tibble: 28 x 3
#>    sex    race  year     
#>    <chr>  <chr> <chr>    
#>  1 Female Black 1981-1985
#>  2 Female Black 1986-1990
#>  3 Female Black 1991-1995
#>  4 Female Black 1996-2000
#>  5 Female Black 2001-2005
#>  6 Female Black 2006-2010
#>  7 Female Black 2011-2015
#>  8 Female Other 1981-1985
#>  9 Female Other 1986-1990
#> 10 Female Other 1991-1995
#> # … with 18 more rows

d_year %>%
  complete_year_groups(sex, race) %>%
  dplyr::arrange(sex, race, year)

#> # A tibble: 28 x 3
#>    sex    race  year     
#>    <chr>  <chr> <chr>    
#>  1 Female Black 1981-1985
#>  2 Female Black 1986-1990
#>  3 Female Black 1991-1995
#>  4 Female Black 1996-2000
#>  5 Female Black 2001-2005
#>  6 Female Black 2006-2010
#>  7 Female Black 2011-2015
#>  8 Female Other 1981-1985
#>  9 Female Other 1986-1990
#> 10 Female Other 1991-1995
#> # … with 18 more rows