SEER Florida population data, expanded Race

Florida county-level population estimates from 1990--2016 in 19 age groups, with 4 expanded races by origin. Downloaded from SEER on 2019-01-24 (see references for direct links). Recoding of column values was performed according to the guidelines available at https://seer.cancer.gov/popdata/popdic.html. The original SEER data includes 19 age groups, but the data set provided by this package includes age "0" with ages "1 - 4" in the "0 - 4" group to match the coding in the FCDS data.

seer_pop_fl_exp_race

Format

A data frame with 334273 rows and 11 variables:

year

Year

state

State postal abbreviation

state_fips

State FIPS code

county_fips

County FIPS code

registry

Registry

race

Race: White, Black, American Indian/Alaska Native, Asian or Pacific Islander

origin

Origin: Non-Hispanic, Hispanic. Not applicable for data prior to 1990.

sex

Sex

age_group

Age: 19 groups from 0, 1--4, 5--9, … 85+

population

Population

county_name

County Name, see county_fips_fl

References

https://seer.cancer.gov/popdata/

https://seer.cancer.gov/popdata/yr1990_2016.19ages/fl.1990_2016.19ages.txt.gz

https://seer.cancer.gov/popdata/popdic.html

See also