Complete Age Groups

Completes age groups by adding missing age groups, either within the age range from age_low to age_high or using the full age list from seer_std_ages. If the columns age_low or age_high are missing from the input data, expand_age_groups() is first called to expand the age group variable.

complete_age_groups(data, age_low = NULL, age_high = NULL, ...,
  age_group = age_group, fill = list(n = 0), include_unknown = FALSE,
  std_age_groups = fcds_const("age_group"))

Arguments

data
A data frame.

age_low
Youngest age (inclusive).

age_high
Eldest age (inclusive).

...
Specification of columns to expand. Columns can be atomic vectors or lists.

To find all unique combinations of x, y and z, including those not found in the data, supply each variable as a separate argument. To find only the combinations that occur in the data, use nest: expand(df, nesting(x, y, z)).

You can combine the two forms. For example, expand(df, nesting(school_id, student_id), date) would produce a row for every student for each date.

For factors, the full set of levels (not just those that appear in the data) are used. For continuous variables, you may need to fill in values that don't appear in the data: to do so use expressions like year = 2010:2020 or year = full_seq(year,1).

Length-zero (empty) elements are automatically dropped.

age_group
Unquoted column name containing the age grouping.

fill
A named list that for each variable supplies a single value to use instead of NA for missing combinations.

include_unknown
Should the "Unknown" age group be included?

std_age_groups
Character vector containing expected (or standard) age groups.

...
Arguments passed on to expand_age_groups

data

A data frame.

age_group

Unquoted column name containing the age grouping.

age_low

Youngest age (inclusive).

age_high

Eldest age (inclusive).

sep

Separator between columns. If character, is interpreted as a regular expression. The default value is a regular expression that matches any sequence of non-alphanumeric values. If numeric, interpreted as positions to split at. Positive values start at 1 at the far-left of the string; negative value start at -1 at the far-right of the string. The length of sep should be one less than into.

See also

Examples

tibble::tibble(
  age_group = c("10 - 14", "15 - 19", "25 - 29"),
  n = 10:12
) %>%
  complete_age_groups(10, 35)

#> # A tibble: 5 x 2
#>   age_group     n
#>   <ord>     <dbl>
#> 1 10 - 14      10
#> 2 15 - 19      11
#> 3 20 - 24       0
#> 4 25 - 29      12
#> 5 30 - 34       0