Count FCDS Cases

Helper function to count cancer cases in the FCDS data by year and age group, in addition to any groups already present in the data. For convenience, you may additionally filter to include particular values of sex, race, year, county name and hispanic ethnicity. See fcds_const() for more information about possible values for these variables. By default, count_fcds() ensures that age_group, year_group, and year are included in the grouping variables if they are present in the data. If they are not, or if they are not present in the FCDS data, then it would be better to use dplyr::count() directly.

count_fcds(data, ..., sex = NULL, race = NULL, origin = NULL,
  moffitt_catchment = FALSE, default_groups = c("year_group", "year",
  "age_group"), discard_unseen_levels = TRUE)

Arguments

data
A data frame

...
Unquoted column names to be added to the grouping of the output and subsequent counting.

sex
Character vector of values of sex to be included in count, or TRUE to include all values of sex present in input data

race
Character vector of values of race to be included in count, or TRUE to include all values of race present in input data

origin
Character vector of values of origin to be included in count, or TRUE to include all values of origin present in input data

moffitt_catchment
Limit counties to those in the catchment area of the Moffitt Cancer Center.

default_groups
Variables that should be included in the grouping, prior to counting cancer cases. Set to NULL to use only the groups already present in the input data.

discard_unseen_levels
If TRUE, drops factor levels that don't appear in the results. If FALSE, then no changes are made to the factor levels. If a character string of column names, then only unobserved levels in those columns are dropped.

Value

A grouped data frame with counts. The output groups includes the union of the groups of the original input data, the groups specified by the columns indicated in ..., and the default_groups added by count_fcds() (modifyable by the default_groups argument).

All factor levels will be modified to include only those levels that appear in the final output across all groups.

Examples

fcds_example %>%
  dplyr::filter(county_name == "Pinellas") %>%
  count_fcds(cancer_site_group, sex = "Male")

#> # A tibble: 285 x 6
#> # Groups:   cancer_site_group, sex, year_group, year, age_group [285]
#>    cancer_site_group       sex   year_group year  age_group     n
#>    <fct>                   <fct> <fct>      <chr> <fct>     <int>
#>  1 Oral Cavity and Pharynx Male  1992-1996  1994  0 - 4         1
#>  2 Oral Cavity and Pharynx Male  1992-1996  1994  55 - 59       1
#>  3 Oral Cavity and Pharynx Male  1997-2001  1999  70 - 74       1
#>  4 Oral Cavity and Pharynx Male  2002-2006  2004  15 - 19       1
#>  5 Oral Cavity and Pharynx Male  2002-2006  2004  75 - 79       1
#>  6 Oral Cavity and Pharynx Male  2007-2011  2009  70 - 74       2
#>  7 Esophagus               Male  1987-1991  1989  70 - 74       1
#>  8 Esophagus               Male  1987-1991  1989  75 - 79       1
#>  9 Esophagus               Male  1997-2001  1999  65 - 69       1
#> 10 Esophagus               Male  2007-2011  2009  50 - 54       1
#> # … with 275 more rows